// backend/src/models/estados.models.ts
export enum NivelDeRiesgo {
    BAJO = 'bajo',
    MODERADO = 'moderado',
    URGENTE = 'urgente'
}

export enum Tema {
    PISO_ROTO = 'piso roto',
    CALLE = 'calle',
    ALUMBRADO = 'alumbrado'
}

export enum Localidad {
    JOSE_C_PAZ = 'Jose C Paz',
    SAN_MIGUEL = 'San Miguel',
    POLVORINES = 'Polvorines'
}

export enum Estado {
    EN_REVISION = 'En Revisión',
    COMPLETADO = 'Completado',
    ANULADO = 'Anulado',
  }
  